Create cloud subscription component
