+++ apps/web/src/app/organizations/billing/organization-subscription.component.html
@@ -50,194 +50,17 @@
       <span>{{ "reinstateSubscription" | i18n }}</span>
     </button>
   </app-callout>
-  <ng-container *ngIf="!selfHosted">
-    <div class="row">
-      <div class="col-4">
-        <dl>
-          <dt>{{ "billingPlan" | i18n }}</dt>
-          <dd>{{ sub.plan.name }}</dd>
-          <ng-container *ngIf="subscription">
-            <dt>{{ "status" | i18n }}</dt>
-            <dd>
-              <span class="text-capitalize">{{
-                isSponsoredSubscription ? "sponsored" : subscription.status || "-"
-              }}</span>
-              <span bitBadge badgeType="warning" *ngIf="subscriptionMarkedForCancel">{{
-                "pendingCancellation" | i18n
-              }}</span>
-            </dd>
-            <dt>{{ "nextCharge" | i18n }}</dt>
-            <dd>
-              {{
-                nextInvoice
-                  ? (nextInvoice.date | date: "mediumDate") +
-                    ", " +
-                    (nextInvoice.amount | currency: "$")
-                  : "-"
-              }}
-            </dd>
-          </ng-container>
-        </dl>
-      </div>
-      <div class="col-8" *ngIf="subscription">
-        <strong class="d-block mb-1">{{ "details" | i18n }}</strong>
-        <table class="table">
-          <tbody>
-            <tr *ngFor="let i of subscription.items">
-              <td>
-                {{ i.name }} {{ i.quantity > 1 ? "&times;" + i.quantity : "" }} @
-                {{ i.amount | currency: "$" }}
-              </td>
-              <td>{{ i.quantity * i.amount | currency: "$" }} /{{ i.interval | i18n }}</td>
-            </tr>
-          </tbody>
-        </table>
-      </div>
-      <ng-container *ngIf="userOrg?.providerId != null">
-        <div class="col-sm">
-          <dl>
-            <dt>{{ "provider" | i18n }}</dt>
-            <dd>{{ "yourProviderIs" | i18n: userOrg.providerName }}</dd>
-          </dl>
-        </div>
-      </ng-container>
-    </div>
-    <ng-container>
-      <button
-        type="button"
-        class="btn btn-outline-secondary"
-        (click)="changePlan()"
-        *ngIf="showChangePlanButton"
-      >
-        {{ "changeBillingPlan" | i18n }}
-      </button>
-      <app-change-plan
-        [organizationId]="organizationId"
-        (onChanged)="closeChangePlan()"
-        (onCanceled)="closeChangePlan()"
-        *ngIf="showChangePlan"
-      ></app-change-plan>
-    </ng-container>
-    <h2 class="spaced-header">{{ "manageSubscription" | i18n }}</h2>
-    <p class="mb-4">{{ subscriptionDesc }}</p>
-    <ng-container
-      *ngIf="
-        subscription && canAdjustSeats && !subscription.cancelled && !subscriptionMarkedForCancel
-      "
-    >
-      <div class="mt-3">
-        <app-adjust-subscription
-          [seatPrice]="seatPrice"
-          [organizationId]="organizationId"
-          [interval]="billingInterval"
-          [currentSeatCount]="seats"
-          [maxAutoscaleSeats]="maxAutoscaleSeats"
-          (onAdjusted)="subscriptionAdjusted()"
-        >
-        </app-adjust-subscription>
-      </div>
-    </ng-container>
-    <button
-      #removeSponsorshipBtn
-      type="button"
-      class="btn btn-outline-danger btn-submit"
-      (click)="removeSponsorship()"
-      [appApiAction]="removeSponsorshipPromise"
-      [disabled]="$any(removeSponsorshipBtn).loading"
-      *ngIf="isSponsoredSubscription"
-    >
-      <i class="bwi bwi-spinner bwi-spin" title="{{ 'loading' | i18n }}" aria-hidden="true"></i>
-      <span>{{ "removeSponsorship" | i18n }}</span>
-    </button>
-    <h2 class="spaced-header">{{ "storage" | i18n }}</h2>
-    <p>{{ "subscriptionStorage" | i18n: sub.maxStorageGb || 0:sub.storageName || "0 MB" }}</p>
-    <div class="progress">
-      <div
-        class="progress-bar bg-success"
-        role="progressbar"
-        [ngStyle]="{ width: storageProgressWidth + '%' }"
-        [attr.aria-valuenow]="storagePercentage"
-        aria-valuemin="0"
-        aria-valuemax="100"
-      >
-        {{ storagePercentage / 100 | percent }}
-      </div>
-    </div>
-    <ng-container *ngIf="subscription && !subscription.cancelled && !subscriptionMarkedForCancel">
-      <div class="mt-3">
-        <div class="d-flex" *ngIf="!showAdjustStorage">
-          <button type="button" class="btn btn-outline-secondary" (click)="adjustStorage(true)">
-            {{ "addStorage" | i18n }}
-          </button>
-          <button
-            type="button"
-            class="btn btn-outline-secondary ml-1"
-            (click)="adjustStorage(false)"
-          >
-            {{ "removeStorage" | i18n }}
-          </button>
-        </div>
-        <app-adjust-storage
-          [storageGbPrice]="storageGbPrice"
-          [add]="adjustStorageAdd"
-          [organizationId]="organizationId"
-          [interval]="billingInterval"
-          (onAdjusted)="closeStorage(true)"
-          (onCanceled)="closeStorage(false)"
-          *ngIf="showAdjustStorage"
-        ></app-adjust-storage>
-      </div>
-    </ng-container>
-    <!--Switch to i18n-->
-    <h2 class="spaced-header">{{ "selfHostingTitle" | i18n }}</h2>
-    <p class="mb-4">
-      {{ "selfHostingEnterpriseOrganizationSectionCopy" | i18n }}
-    </p>
-    <div class="d-flex">
-      <button
-        type="button"
-        class="btn btn-outline-secondary"
-        (click)="downloadLicense()"
-        *ngIf="canDownloadLicense"
-        [disabled]="showDownloadLicense"
-      >
-        {{ "downloadLicense" | i18n }}
-      </button>
-      <button
-        type="button"
-        class="btn btn-outline-secondary ml-1"
-        (click)="manageBillingSync()"
-        *ngIf="canManageBillingSync"
-      >
-        {{ (hasBillingSyncToken ? "manageBillingSync" : "setUpBillingSync") | i18n }}
-      </button>
-    </div>
-    <div class="mt-3" *ngIf="showDownloadLicense">
-      <app-download-license
-        [organizationId]="organizationId"
-        (onDownloaded)="closeDownloadLicense()"
-        (onCanceled)="closeDownloadLicense()"
-      ></app-download-license>
-    </div>
-    <h2 class="spaced-header">{{ "additionalOptions" | i18n }}</h2>
-    <p class="mb-4">
-      {{ "additionalOptionsDesc" | i18n }}
-    </p>
-    <div class="d-flex">
-      <button
-        #cancelBtn
-        type="button"
-        class="btn btn-outline-danger btn-submit ml-1"
-        (click)="cancel()"
-        [appApiAction]="cancelPromise"
-        [disabled]="$any(cancelBtn).loading"
-        *ngIf="subscription && !subscription.cancelled && !subscriptionMarkedForCancel"
-      >
-        <i class="bwi bwi-spinner bwi-spin" title="{{ 'loading' | i18n }}" aria-hidden="true"></i>
-        <span>{{ "cancelSubscription" | i18n }}</span>
-      </button>
-    </div>
-  </ng-container>
+  <app-org-subscription-cloud
+    *ngIf="!selfHosted && !loading"
+    [sub]="sub"
+    [organizationId]="organizationId"
+    [loading]="loading"
+    [hasBillingSyncToken]="hasBillingSyncToken"
+    [subscriptionMarkedForCancel]="subscriptionMarkedForCancel"
+    [userOrg]="userOrg"
+    (reload)="load()"
+  >
+  </app-org-subscription-cloud>
   <app-org-subscription-selfhost
     *ngIf="selfHosted && !loading"
     [sub]="sub"
@@ -246,4 +69,3 @@
     (reload)="load()"
   ></app-org-subscription-selfhost>
 </ng-container>
-<ng-template #setupBillingSyncTemplate></ng-template>
